requirements.txt
aniso8601==9.0.1
attrs==23.1.0
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
flask-restx==1.1.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
pkgutil_resolve_name==1.3.10
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
Werkzeug==2.2.3
zipp==3.15.0
